claim_processing_automation/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1

# PDF Processing & OCR
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
fitz==0.0.1.dev2
Pillow==10.1.0

# LLM & AI Processing
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.8
anthropic==0.7.8
transformers==4.36.0
torch==2.1.1

# Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Web Scraping & HTTP
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Payment Processing
stripe==7.8.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.8

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Background Tasks
celery==5.3.4
kombu==5.3.4

# Cloud Storage
boto3==1.34.0
google-cloud-storage==2.10.0

# Email
sendgrid==6.11.0